Add summary stats for hashtags
This commit improves the root summary stats route to allow summary
statistics for hashtags tracked by the system. The stats are 'number of
users', 'number of changesets', 'km of roads edited', 'number of
buildings edited', 'number of edits' and 'latest edit' for that hashtag.

The root route file is renamed summary.js to reflect the new
functionality.
